platform/x86: Intel PMT Crashlog capability driver
commit 5ef9998c96b0c99c49c202054586967e609286d2 upstream.

Backport summary: backport to kernel 5.10.18 for intel PMT (Platform
Monitoring Technology) support

Add support for the Intel Platform Monitoring Technology crashlog
interface. This interface provides a few sysfs values to allow for
controlling the crashlog telemetry interface as well as a character
driver to allow for mapping the crashlog memory region so that it can be
accessed after a crashlog has been recorded.

This driver is meant to only support the server version of the crashlog
which is identified as crash_type 1 with a version of zero. Currently no
other types are supported.

Signed-off-by: Alexander Duyck <alexander.h.duyck@linux.intel.com>
Signed-off-by: David E. Box <david.e.box@linux.intel.com>
Reviewed-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
Signed-off-by: yingbao jia<yingbao.jia@intel.com>
Signed-off-by: Fam Zheng <fam.zheng@bytedance.com>