KVM: X86: Reset the vcpu->run->flags at the beginning of vcpu_run
commit 15aad3be9adb3fb7fba84190a2ce57d66e8b51da upstream.

Reset the vcpu->run->flags at the beginning of kvm_arch_vcpu_ioctl_run.
It can avoid every thunk of code that needs to set the flag clear it,
which increases the odds of missing a case and ending up with a flag in
an undefined state.

Intel-SIG: commit 15aad3be9adb KVM: X86: Reset the vcpu->run->flags at the beginning of vcpu_run
Backport to add KVM support for Bus Lock VM Exit

Signed-off-by: Chenyi Qiang <chenyi.qiang@intel.com>
Message-Id: <20201106090315.18606-3-chenyi.qiang@intel.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
[ Aichun Shi: resolve conflict and amend commit log ]
Signed-off-by: Aichun Shi <aichun.shi@intel.com>
Signed-off-by: Jiawei Gu <gujiawei.will@bytedance.com>