PCI/ERR: Rename reset_link() to reset_subordinates()
commit 8f1bbfbc3596d401b60d1562b27ec28c2724f60d upstream
Backport summary: Backport to kernel 5.10.18 to enable PCIe/RCEC.

reset_link() appears to be misnamed.  The point is to reset any devices
below a given bridge, so rename it to reset_subordinates() to make it clear
that we are passing a bridge with the intent to reset the devices below it.

Link: https://lore.kernel.org/r/20201121001036.8560-5-sean.v.kelley@intel.com
Tested-by: Jonathan Cameron <Jonathan.Cameron@huawei.com> # non-native/no RCEC
Signed-off-by: Sean V Kelley <sean.v.kelley@intel.com>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
Reviewed-by: Kuppuswamy Sathyanarayanan <sathyanarayanan.kuppuswamy@linux.intel.com>
Acked-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
Signed-off-by: Youquan Song <youquan.song@intel.com>
Signed-off-by: Fam Zheng <fam.zheng@bytedance.com>