ovl: do not get metacopy for userxattr
commit 87b2c60c6127eaa3b33f3e4c4166f6af80e1aaa9 upstream.

When looking up an inode on the lower layer for which the mounter lacks
read permisison the metacopy check will fail.  This causes the lookup to
fail as well, even though the directory is readable.

So ignore EACCES for the "userxattr" case and assume no metacopy for the
unreadable file.

Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Signed-off-by: Xin Yin <yinxin.x@bytedance.com>