KVM: x86/mmu: Use an rwlock for the x86 MMU
commit 531810caa9f4bc99ffbb90e09256792c56a6b07a upstream.

backport notes:

1. Dirty ring implementation
fb04a1eddb1a65b6588a021bdc132270d5ae48bb has not been introduced yet, so
abandon the original change in dirty_ring.c, which doesn't exist in 5.10
code base.

2. This commit should be eariler than upstream commit
4a38162ee9f10f5f67c36f4f5aa4f6be2657efd5. But it has been ported to this
code base and uses spinlock instead of rwlock. So take this chance to
turn it back to rwlock with rwlock introduction.

Add a read / write lock to be used in place of the MMU spinlock on x86.
The rwlock will enable the TDP MMU to handle page faults, and other
operations in parallel in future commits.

Reviewed-by: Peter Feiner <pfeiner@google.com>
Signed-off-by: Ben Gardon <bgardon@google.com>

Message-Id: <20210202185734.1680553-19-bgardon@google.com>
[Introduce virt/kvm/mmu_lock.h - Paolo]
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Jiawei Gu <gujiawei.will@bytedance.com>