x86/fault: Don't run fixups for SMAP violations
commit ca247283781d754216395a41c5e8be8ec79a5f1c upstream.

A SMAP-violating kernel access is not a recoverable condition.  Imagine
kernel code that, outside of a uaccess region, dereferences a pointer to
the user range by accident.  If SMAP is on, this will reliably generate
as an intentional user access.  This makes it easy for bugs to be
overlooked if code is inadequately tested both with and without SMAP.

This was discovered because BPF can generate invalid accesses to user
memory, but those warnings only got printed if SMAP was off. Make it so
that this type of error will be discovered with SMAP on as well.

 [ bp: Massage commit message. ]

Intel-SIG: commit ca247283781d x86/fault: Don't run fixups for SMAP violations.
Backport summary: This is preparation patch for CET.

Signed-off-by: Andy Lutomirski <luto@kernel.org>
Signed-off-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/66a02343624b1ff46f02a838c497fc05c1a871b3.1612924255.git.luto@kernel.org
[ Jing Han amend commit log ]
Signed-off-by: Jing Han <jing1.han@intel.com>
Signed-off-by: Gang Li <ligang.bdlg@bytedance.com>