intel_idle: update ICX C6 data
commit d484b8bfc6fa71a088e4ac85d9ce11aa0385867e upstream.

Change IceLake Xeon C6 latency from 128 us to 170 us. The latency
was measured with the "wult" tool and corresponds to the 99.99th
percentile when measuring with the "nic" method. Note, the 128 us
figure correspond to the median latency, but in intel_idle we use
the "worst case" latency figure instead.

C6 target residency was increased from 384 us to 600 us, which may
result in less C6 residency in some workloads. This value was tested
and compared to values 384, and 1000. Value 600 is a reasonable
tradeoff between power and performance.

Intel-SIG: commit d484b8bfc6fa intel_idle: update ICX C6 data.
Backport intel idle driver for 5.10

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>
Acked-by: Zhang Rui <rui.zhang@intel.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
[ yingbao jia: amend commit log ]
Signed-off-by: yingbao jia <yingbao.jia@intel.com>
Signed-off-by: Gang Li <ligang.bdlg@bytedance.com>