x86/fpu: Reject invalid MXCSR values in copy_kernel_to_xstate()
commit 947f4947cf00ea1e6d319eb182c64ea51ba4de8d upstream.

Instead of masking out reserved bits, check them and reject the provided
state as invalid if not zero.

Intel-SIG: commit 947f4947cf00 x86/fpu: Reject invalid MXCSR values in
copy_kernel_to_xstate().
Intel Advanced Matrix Extensions(AMX) is one of the key features on Sapphire
Rapids(SPR) and it depends on xsave, this patch is part of Thomas Gleixner's
x86 FPU code cleanup that upstream AMX implementation is building on, and is
also key to all code for being developed in upstream that depends on xsave.

Suggested-by: Andy Lutomirski <luto@kernel.org>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Borislav Petkov <bp@suse.de>
Reviewed-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/20210623121452.308388343@linutronix.de
[ Lin Wang: amend commit log ]
Signed-off-by: Lin Wang <lin.x.wang@intel.com>
Signed-off-by: Gang Li <ligang.bdlg@bytedance.com>