x86/fpu/regset: Move fpu__read_begin() into regset
commit 5a32fac8dbe8adc08c10e2c8770c95aebfc627cd upstream.

The function can only be used from the regset get() callbacks safely. So
there is no reason to have it globally exposed.

Intel-SIG: commit 5a32fac8dbe8 x86/fpu/regset: Move fpu__read_begin() into
regset.
Intel Advanced Matrix Extensions(AMX) is one of the key features on Sapphire
Rapids(SPR) and it depends on xsave, this patch is part of Thomas Gleixner's
x86 FPU code cleanup that upstream AMX implementation is building on, and is
also key to all code for being developed in upstream that depends on xsave.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Borislav Petkov <bp@suse.de>
Reviewed-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/20210623121453.234942936@linutronix.de
[ Lin Wang: amend commit log ]
Signed-off-by: Lin Wang <lin.x.wang@intel.com>
Signed-off-by: Gang Li <ligang.bdlg@bytedance.com>