perf/x86/intel/uncore: Add Sapphire Rapids server IMC support
commit 85f2e30f987ecc73fbb5e24eda0f36ba7f337c5c upstream.

The Sapphire Rapids IMC provides the interface to the DRAM and
communicates to the rest of the uncore through the M2M block.

The layout of the control registers for a IMC uncore unit is a little
bit different from the generic one. There is a fixed counter for IMC.
So a specific format and ops are required. Expose the common MMIO ops
which can be reused.

Intel-SIG: commit 85f2e30f987e perf/x86/intel/uncore: Add Sapphire
Rapids server IMC support.
Backport SPR uncore and free-running counter support

Signed-off-by: Kan Liang <kan.liang@linux.intel.com>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Reviewed-by: Andi Kleen <ak@linux.intel.com>
Link: https://lore.kernel.org/r/1625087320-194204-8-git-send-email-kan.liang@linux.intel.com
[ Yunying Sun amend commit log ]
Signed-off-by: Yunying Sun <yunying.sun@intel.com>
Signed-off-by: Gang Li <ligang.bdlg@bytedance.com>