clocksource/drivers/arch_arm_timer: Move workaround synchronisation around
commit db26f8f2da92471e9f7f71ec78d6fa455cd9c821 upstream.

We currently handle synchronisation when workarounds are enabled
by having an ISB in the __arch_counter_get_cnt?ct_stable() helpers.

While this works, this prevents us from relaxing this synchronisation.

Instead, move it closer to the point where the synchronisation is
actually needed. Further patches will subsequently relax this.

Signed-off-by: Marc Zyngier <maz@kernel.org>
Link: https://lore.kernel.org/r/20211017124225.3018098-14-maz@kernel.org
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Signed-off-by: chenyi <chenyi211@huawei.com>
Signed-off-by: Jinhui Guo <guojinhui.liam@bytedance.com>