bytedance: checkpatch: ignore checking format of "[ Upstream commit $id ]"
There are a lot of patches which need to be backported to our own
branch, some patches with a string like "[ Upstream commit $id ]"
to the front of the commit log. But checkpatch always complains
about this and our CI/CD always runs checkpatch before test. In
order to make this test step useful, we ignore the check of the
format like "[ Upstream commit $id ]".

Signed-off-by: Muchun Song <songmuchun@bytedance.com>