net: allow gro_max_size to exceed 65536
commit 0fe79f28bfaf73b66b7b1562d2468f94aa03bd12 upstream.

Allow the gro_max_size to exceed a value larger than 65536.

There weren't really any external limitations that prevented this other
than the fact that IPv4 only supports a 16 bit length field. Since we have
the option of adding a hop-by-hop header for IPv6 we can allow IPv6 to
exceed this value and for IPv4 and non-TCP flows we can cap things at 65536
via a constant rather than relying on gro_max_size.

[edumazet] limit GRO_MAX_SIZE to (8 * 65535) to avoid overflows.

Signed-off-by: Alexander Duyck <alexanderduyck@fb.com>
Signed-off-by: Eric Dumazet <edumazet@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

 Conflicts:
	drivers/net/ethernet/mellanox/mlx5/core/en_rx.c
  Ignored the change in en/rx.c for function
  mlx5e_hw_gro_skb_has_enough_space, which is used by
  mlx5e_handle_rx_cqe_mpwrq_shampo and it is not present in our code base.

Signed-off-by: A K M Fazla Mehrab <a.mehrab@bytedance.com>